+++ runtime/index.js
@@ -17,4 +17,5 @@ export * from "./anim.js";
 export * from "./terminal.js";
 export * from "./nav.js";
 export * from "./lsnav.js";
+export * from "./diagramzoom.js";
 export * from "./tooltip.js";